pyproject.toml
[tool.poetry]
name = "django_template_project"
version = "0.1.0"
description = ""
authors = ["vildan-valeev <vildan.maydanovich.valeev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.2"
django-cors-headers = "^3.11.0"
djangorestframework = "^3.13.1"
drf-yasg = "^1.20.0"
django-dramatiq = "^0.10.0"
django-split-settings = "^1.1.0"
environs = "^9.5.0"
redis = "^4.1.2"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
